requirements.txt
fastapi==0.105.0
uvicorn==0.23.2
SQLAlchemy==2.0.32
pydantic==1.10.16
email-validator==1.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.8
